Create a new owning document for function callbacks which fixes rt71345. Added the two test cases from that ticket to unit tests.
Discussion: LibXSLT assumed the owner of the XPath node was the document being transformed. If the document was actually created by an intermediate extension e.g. exslt:node-set it won't be owned by us. This change creates an entirely new document for every callback that is passed to perl_dispatch and lasts for the duration of the dispatch or the parent scope if it or any owned elements are referenced.